Build release using docker(-desktop) to provide a stable and appropriate OS to work from.
The bulk of this was stolen.... er... plagerized directly from the
pi-gen scripts and uses almost all of the same logic. I've done
some customizing to do things such as provide a basic config file
with information in it so that the next person can figure out what
it's all about.

The '*.secret' file is designed to allow the config file to be
distributed without disclosing anything you don't want/need to give
away.

Offered back to the community.